refactor(setPath): divide code in setPath in different files to make it cleaner
