+++ packages/shared/config/env-vars/index.js
@@ -1,0 +1,5 @@
+const node = require('./node');
+
+module.exports = {
+  node,
+};
