+++ packages/shared/config/env-vars/node.js
@@ -1,0 +1,3 @@
+module.exports = {
+  env: process.env.NODE_ENV,
+};
