--- packages/shared/config/environment.js
@@ -1,3 +1,0 @@
-module.exports = {
-  NODE_ENV: process.env.NODE_ENV,
-};
