+++ packages/shared/src/utils/dotenv/index.js
@@ -1,0 +1,6 @@
+// Internal dependendencies
+const setPath = require('./set-path');
+
+module.exports = {
+  setPath,
+};
