+++ packages/shared/src/utils/env-checkers/index.js
@@ -1,0 +1,8 @@
+// Internal dependencies
+const isDevEnvironment = require('./is-development');
+const isProdEnvironment = require('./is-production');
+
+module.exports = {
+  isDevEnvironment,
+  isProdEnvironment,
+};
