+++ packages/shared/src/utils/env-checkers/is-production.js
@@ -1,0 +1,12 @@
+// Internal dependencies
+const softwareEnvironments = require('../constants/software-environment');
+const { envVars } = require('../../../config');
+
+const isProdEnvironment = () => {
+  if (envVars.node.env === softwareEnvironments.PROD) {
+    return true;
+  }
+  return false;
+};
+
+module.exports = isProdEnvironment;
